+++ index.js
@@ -1,5 +1,4 @@
-const expect = require('chai').expect;
-var toStringRef = require("./mapping");
+const toStringRef = require("./mapping");
 
 const MajorColors = [
     "white",
@@ -43,27 +42,5 @@ function GetPairNumberFromColors(majorColor, minorColor) {
     return majorIndex * MinorColors.length + minorIndex + 1;
 }
 
-function testNumberToPair(number, expectedMajor, expectedMinor) {
-    const pairOfColors = GetColorFromPairNumber(number);
-    console.log(`${number} = ${expectedMajor} ${expectedMinor}`);
-    expect(pairOfColors.major).equals(expectedMajor);
-    expect(pairOfColors.minor).equals(expectedMinor);
-}
-
-function testColorToNumber(majorColor, minorColor, expectedNumber) {
-    const pairNumber = GetPairNumberFromColors(majorColor, minorColor);
-    console.log(`${majorColor} ${minorColor} = ${pairNumber}`);
-    expect(pairNumber).to.equal(expectedNumber);
-}
-
-console.log("color coder");
-testNumberToPair(4, "white", "brown");
-testNumberToPair(25, "violet", "slate");
-testColorToNumber("black", "brown", 14);
-testColorToNumber("yellow", "green", 18);
-
 toStringRef();
-
-module.exports= GetPairNumberFromColors;
-module.exports= MajorColors;
-module.exports= MinorColors;
+module.exports= {GetPairNumberFromColors, GetColorFromPairNumber, MajorColors, MinorColors};
